Modified gmvault-runner.sh and gdrive-backup-runner.sh to abort and print a useful message if they cannot succeed due to the target volume not being mounted

+++ bin/gmvault-runner.sh
@@ -20,14 +20,19 @@ function cleanup() {
 trap cleanup EXIT
 
 
+if [ ! -d $GMVAULT_DB ]; then
+  echo -e "$(date): GMVAULT skipped backup of $ACCOUNT due to $GMVAULT_DB not existing or not being mounted correctly"
+  trap - EXIT
+  exit 1
+fi
 $GMVAULT_EXE sync "$ACCOUNT" -d "$GMVAULT_DB" >> "$tmplogfile" 2>&1
 gmvaultstatus=$?
 if [ $gmvaultstatus -eq 0 ]; then
-  echo -e "GMVAULT backed up $ACCOUNT to $GMVAULT_DB successfully on $(date)" | tee -a $LOGFILE
+  echo -e "$(date): GMVAULT backed up $ACCOUNT to $GMVAULT_DB successfully" | tee -a $LOGFILE
   rm "$tmplogfile"
   trap - EXIT
 else
-  echo -e "GMVAULT FAILED to backp $ACCOUNT to $GMVAULT_DB on $(date); gmvault log stored at $tmplogfile"
+  echo -e "$(date): GMVAULT FAILED to backup $ACCOUNT to $GMVAULT_DB; gmvault log stored at $tmplogfile"
   trap - EXIT
 fi
